Organisation
Split project and annotation class into separate files.
Move ui and test folders into the project root.
