+++ test/test_about_dialog.py
