+++ test/test_new_project_dialog.py
