+++ ui/shortcut_settings_dialog.ui
